Collision Detection and Centering of Fixtures
Added the ability for the joystick rectangle to detect collision.
In the future this will be used to create a "selection mode", in which the
fixture to be controlled can be selected first.

Additionally added an option to specify the number of fixtures needed (still in the code, not yet on screen)
and make them align with even spacing around the center.