Added calibration_mode and general button support
Added an initialization function in the calibration class. Changed the "Run" file so that pressing
the 2nd and 3rd (start at 0) button of the main Joystick will trigger the initialization and run a calibration_mode function,
which allows the user to calibrate the fixture

Also added an universal support for all of the buttons on the joystick.
The standard button implementation through pyglet will
always give an output as long as the button is pressed.
This implementation now also allows for single-triggering events with a button press.
Useful for future features.